add grpc dependency

+++ services/product/src/main/java/org/microservices/product/DTO/ProductPurchaseResponse.java
@@ -6,7 +6,7 @@ public record ProductPurchaseResponse(
         Integer productId,
         String name,
         String description,
-        BigDecimal price,
+        double price,
         int quantity
 ) {
 }
